split up code.  Add Client disconnect handler
