+++ ascii_telnet_server/ascii_server.py
@@ -1,0 +1,72 @@
+# coding=utf-8
+# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+#  Copyright (c) 2008, Martin W. Kirst All rights reserved.
+#
+#  Redistribution and use in source and binary forms, with or without
+#  modification, are permitted provided that the following conditions are
+#  met:
+#
+#  Redistributions of source code must retain the above copyright notice,
+#  this list of conditions and the following disclaimer.
+#  Redistributions in binary form must reproduce the above copyright notice,
+#  this list of conditions and the following disclaimer in the documentation
+#  and/or other materials provided with the distribution.
+#
+#  THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS
+#  IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED
+#  TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A
+#  PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+#  HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+#  SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED
+#  TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
+#  PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
+#  LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
+#  NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+#  SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+from __future__ import division, print_function
+
+import errno
+import socket
+
+from ascii_telnet_server.ascii_movie import Movie
+from ascii_telnet_server.ascii_player import VT100Player
+
+try:
+    # noinspection PyCompatibility
+    from socketserver import StreamRequestHandler, ThreadingMixIn, TCPServer
+except ImportError:  # Py2
+    # noinspection PyCompatibility,PyUnresolvedReferences
+    from SocketServer import StreamRequestHandler, ThreadingMixIn, TCPServer
+
+
+class ThreadedTCPServer(ThreadingMixIn, TCPServer):
+    daemon_threads = True
+
+
+class TelnetRequestHandler(StreamRequestHandler):
+    """
+    Request handler used for multi threaded TCP server
+    @see: SocketServer.StreamRequestHandler
+    """
+
+    filename = None  # filename is set once, so it's immutable and safe for multi threading
+
+    def handle(self):
+        movie = Movie(80, 24)
+        movie.loadMovie(TelnetRequestHandler.filename)
+
+        self.player = VT100Player(movie)
+        self.player.onNextFrame = self.onNextFrame
+        self.player.play()
+
+    def onNextFrame(self, screen_buffer):
+        """
+        Gets the current screen buffer and writes it to the socket.
+        """
+        try:
+            self.wfile.write(screen_buffer.read())
+        except socket.error as e:
+            if e.errno == errno.EPIPE:
+                print("Client Disconnected.")
+                self.player.stop()
